# [ClangImporter] Support lazy member loading for import-as-member globals in extensions