In #71027, I addressed a cycle that could occur when loading import-as-member globals into an extension. However, that fix was minimal—it simply broke the cycle after it occurred. This PR is a more comprehensive fix that eliminates the cycle entirely by making import-as-member importing lazier.

The cycle was caused by the compiler forcing the extension to load its members immediately; it did that because the selective lazy member loading path (the `loadNamedMembers()` method) would crash for these specific extensions because they didn't have a clang node. This PR modifies `loadNamedMembers()` to work correctly, using the context parameter (instead of the absent clang node) to determine the submodule and skipping over lookup for the clang node's members. That allows us to modify name lookup to avoid forcing import-as-members extensions early. In addition to simplifying lazy member loader use sites, I expect this to reduce the number of members imported by ClangImporter, especially in code that uses `NSNotification.Name`, which uses this mechanism for a *massive* number of members.

The first commit also refactors the two code paths that add extension members to the member lookup table so that they share most of their code. It appears to be very difficult to write tests that are guaranteed to hit both of these, so in lieu of doing that, we can at least combine them to increase our confidence that both will work even if only one has been tested. This commit fixes rdar://121479725.