@beccadax 2024-02-01
@swift-ci Please test compiler performance