@beccadax 2024-02-01
If this does actually let the compiler do less work, I'd expect to see:

* Higher `NumUnloadedLazyIterableDeclContexts`
* Higher `NamedLazyMemberLoadSuccessCount`