@beccadax 2024-02-05
Here's the comment it didn't post:

# Summary for main full

**Regressions found (see below)**

- [Debug-batch](#debug-batch-1c2b4fe3aad41d92)
    - [debug-batch brief](#debug-batch-brief-1c2b4fe3aad41d92)
    - [debug-batch detailed](#debug-batch-detailed-1c2b4fe3aad41d92)
- [Release](#release-1c2b4fe3aad41d92)
    - [release brief](#release-brief-1c2b4fe3aad41d92)
    - [release detailed](#release-detailed-1c2b4fe3aad41d92)



<a name="debug-batch-1c2b4fe3aad41d92">

# Debug-batch

<a name="debug-batch-brief-1c2b4fe3aad41d92">

## debug-batch brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (2)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 131,855,857,097,291 | 131,115,941,672,410 | -739,915,424,880 | -0.56%
LLVM.NumLLVMBytesOutput | 3,009,737,544 | 3,013,326,638 | 3,589,094 | 0.12%
</details>

<a name="debug-batch-detailed-1c2b4fe3aad41d92">

## debug-batch detailed
<details>
<summary>Regressed (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Sema.LazyStoragePropertyRequest | 2,094 | 2,122 | 28 | 1.34% :no_entry:
</details>
<details>
<summary>Improved (3)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.NumTotalClangImportedEntities | 1,861,848 | 1,693,767 | -168,080 | -9.03% :white_check_mark:
Sema.HasMissingDesignatedInitializersRequest | 33,971 | 32,893 | -1,078 | -3.17% :white_check_mark:
Sema.InheritsSuperclassInitializersRequest | 42,906 | 41,408 | -1,498 | -3.49% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (349)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ImportSetCacheHit | 4,834,890 | 4,839,709 | 4,818 | 0.1%
AST.ImportSetCacheMiss | 911,288 | 912,273 | 985 | 0.11%
AST.ImportSetFoldHit | 434,505 | 435,348 | 842 | 0.19%
AST.ImportSetFoldMiss | 476,783 | 476,925 | 142 | 0.03%
AST.ModuleShadowCacheHit | 86,482 | 86,223 | -259 | -0.3%
AST.ModuleShadowCacheMiss | 6,663 | 6,651 | -12 | -0.18%
AST.ModuleVisibilityCacheHit | 106,917,078 | 106,971,747 | 54,668 | 0.05%
AST.ModuleVisibilityCacheMiss | 469,619 | 469,757 | 138 | 0.03%
AST.NumASTBytesAllocated | 75,493,778,883 | 75,180,515,965 | -313,262,918 | -0.41%
AST.NumASTScopeExpansions | 10,572,473 | 10,586,844 | 14,371 | 0.14%
AST.NumASTScopeLookups | 11,491,333 | 11,506,213 | 14,880 | 0.13%
AST.NumDecls | 381,782 | 382,012 | 230 | 0.06%
AST.NumDependencies | 326,276 | 326,367 | 91 | 0.03%
AST.NumIncrementalDependencies | 40,148 | 40,148 | 0 | 0.0%
AST.NumLinkLibraries | 0 | 0 | 0 | 0.0%
AST.NumLoadedModules | 398,013 | 398,113 | 100 | 0.03%
AST.NumLocalTypeDecls | 6,784 | 6,785 | 1 | 0.01%
AST.NumMacroPluginDependencies | 36 | 36 | 0 | 0.0%
AST.NumModuleLookupClassMember | 7,231 | 7,196 | -35 | -0.49%
AST.NumModuleLookupValue | 133,243,263 | 133,384,147 | 140,883 | 0.11%
AST.NumObjCMethods | 91,839 | 91,949 | 110 | 0.12%
AST.NumOperators | 922 | 924 | 2 | 0.22%
AST.NumPrecedenceGroups | 176 | 176 | 0 | 0.0%
AST.NumReferencedDynamicNames | 153 | 152 | 0 | -0.33%
AST.NumReferencedMemberNames | 17,092,451 | 17,081,037 | -11,414 | -0.07%
AST.NumReferencedTopLevelNames | 1,478,890 | 1,481,993 | 3,103 | 0.21%
AST.NumSourceBuffers | 451,336 | 451,969 | 632 | 0.14%
AST.NumSourceLines | 14,712,293 | 14,726,177 | 13,884 | 0.09%
AST.NumSourceLinesPerSecond | 6,259,428 | 6,229,570 | -29,858 | -0.48%
Frontend.MaxMallocUsage | 1,044,622,874,440 | 1,035,048,361,024 | -9,574,513,416 | -0.92%
Frontend.NumInstructionsExecuted | 131,855,857,097,291 | 131,115,941,672,410 | -739,915,424,880 | -0.56%
Frontend.NumProcessFailures | 47 | 47 | 0 | 0.0%
IRGen.IRGenRequest | 31,788 | 31,839 | 51 | 0.16%
IRGen.OptimizedIRRequest | 0 | 0 | 0 | 0.0%
IRGen.SymbolObjectCodeRequest | 0 | 0 | 0 | 0.0%
IRModule.NumGOTEntries | 357,061 | 357,343 | 281 | 0.08%
IRModule.NumIRAliases | 306,515 | 306,705 | 189 | 0.06%
IRModule.NumIRBasicBlocks | 10,430,512 | 10,439,647 | 9,135 | 0.09%
IRModule.NumIRComdatSymbols | 0 | 0 | 0 | 0.0%
IRModule.NumIRFunctions | 5,365,693 | 5,372,101 | 6,408 | 0.12%
IRModule.NumIRGlobals | 4,946,823 | 4,952,483 | 5,659 | 0.11%
IRModule.NumIRIFuncs | 0 | 0 | 0 | 0.0%
IRModule.NumIRInsts | 118,667,854 | 118,745,411 | 77,557 | 0.07%
IRModule.NumIRNamedMetaData | 128,672 | 128,878 | 206 | 0.16%
IRModule.NumIRValueSymbols | 10,424,618 | 10,436,759 | 12,141 | 0.12%
LLVM.NumLLVMBytesOutput | 3,009,737,544 | 3,013,326,638 | 3,589,094 | 0.12%
Parse.ExportedSourceFileRequest | 478 | 478 | 0 | 0.0%
Parse.IDEInspectionSecondPassRequest | 0 | 0 | 0 | 0.0%
Parse.NumFunctionsParsed | 756,778 | 757,402 | 624 | 0.08%
Parse.NumIterableDeclContextParsed | 957,762 | 958,828 | 1,065 | 0.11%
Parse.ParseAbstractFunctionBodyRequest | 899,087 | 899,387 | 300 | 0.03%
Parse.ParseMembersRequest | 739,969 | 740,916 | 947 | 0.13%
Parse.ParseSourceFileRequest | 404,252 | 404,867 | 615 | 0.15%
Parse.ParseTopLevelDeclsRequest | 403,574 | 404,189 | 615 | 0.15%
SILGen.ASTLoweringRequest | 33,118 | 33,169 | 51 | 0.16%
SILGen.ParseSILModuleRequest | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenFunctions | 4,241,972 | 4,245,422 | 3,450 | 0.08%
SILModule.NumSILGenGlobalVariables | 147,839 | 147,920 | 81 | 0.05%
SILModule.NumSILGenVtables | 41,773 | 41,792 | 19 | 0.05%
SILModule.NumSILGenWitnessTables | 286,524 | 286,608 | 84 | 0.03%
SILModule.NumSILOptDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILOptFunctions | 4,749,091 | 4,752,909 | 3,818 | 0.08%
SILModule.NumSILOptGlobalVariables | 149,953 | 150,038 | 85 | 0.06%
SILModule.NumSILOptVtables | 42,394 | 42,413 | 19 | 0.04%
SILModule.NumSILOptWitnessTables | 328,344 | 328,467 | 122 | 0.04%
SILOptimizer.ExecuteSILPipelineRequest | 162,958 | 163,215 | 257 | 0.16%
SILOptimizer.LoweredSILRequest | 0 | 0 | 0 | 0.0%
Sema.ABIMembersRequest | 370,331 | 370,554 | 222 | 0.06%
Sema.AbstractGenericSignatureRequest | 132,224 | 132,375 | 151 | 0.11%
Sema.AccessLevelRequest | 23,296,080 | 23,280,506 | -15,574 | -0.07%
Sema.ActorIsolationRequest | 50,263,857 | 50,229,419 | -34,438 | -0.07%
Sema.AllMembersRequest | 205,210 | 205,261 | 51 | 0.02%
Sema.AnyObjectLookupRequest | 225 | 223 | -1 | -0.67%
Sema.ApplyAccessNoteRequest | 5,386,157 | 5,390,407 | 4,250 | 0.08%
Sema.AreAllStoredPropertiesDefaultInitableRequest | 54,724 | 54,753 | 29 | 0.05%
Sema.AssociatedConformanceRequest | 207,862 | 207,860 | -1 | -0.0%
Sema.AttachedPropertyWrapperTypeRequest | 40,331 | 40,338 | 7 | 0.02%
Sema.AttachedPropertyWrappersRequest | 34,784,472 | 34,768,460 | -16,012 | -0.05%
Sema.AttachedResultBuilderRequest | 2,304,722 | 2,306,672 | 1,949 | 0.08%
Sema.BodyInitKindRequest | 99,691 | 99,779 | 88 | 0.09%
Sema.BreakTargetRequest | 8,209 | 8,210 | 1 | 0.01%
Sema.CXXNamespaceMemberLookup | 0 | 0 | 0 | 0.0%
Sema.CallerSideDefaultArgExprRequest | 538,019 | 538,647 | 628 | 0.12%
Sema.CheckDistributedFunctionRequest | 80 | 80 | 0 | 0.0%
Sema.CheckInconsistentAccessLevelOnImport | 63,718 | 63,769 | 51 | 0.08%
Sema.CheckInconsistentAccessLevelOnImportSameFileRequest | 63,956 | 64,007 | 51 | 0.08%
Sema.CheckInconsistentImplementationOnlyImportsRequest | 11,731 | 11,735 | 4 | 0.03%
Sema.CheckInconsistentSPIOnlyImportsRequest | 63,956 | 64,007 | 51 | 0.08%
Sema.CheckInconsistentWeakLinkedImportsRequest | 11,731 | 11,735 | 4 | 0.03%
Sema.CheckRedeclarationRequest | 2,900,813 | 2,902,657 | 1,844 | 0.06%
Sema.ClangCategoryLookupRequest | 0 | 0 | 0 | 0.0%
Sema.ClangDirectLookupRequest | 19,378 | 19,383 | 5 | 0.03%
Sema.ClangRecordMemberLookup | 19,378 | 19,383 | 5 | 0.03%
Sema.ClassAncestryFlagsRequest | 160,701 | 160,945 | 244 | 0.15%
Sema.ClosureEffectsRequest | 170,423 | 170,640 | 217 | 0.13%
Sema.ClosureHasExplicitResultRequest | 64,955 | 65,029 | 74 | 0.11%
Sema.CollectExistentialConformancesRequest | 169,144 | 169,420 | 276 | 0.16%
Sema.CollectOverriddenDeclsRequest | 19,016,008 | 19,001,091 | -14,917 | -0.08%
Sema.CompareDeclSpecializationRequest | 999,387 | 1,000,967 | 1,580 | 0.16%
Sema.CompilerPluginLoadRequest | 36 | 36 | 0 | 0.0%
Sema.ConditionalRequirementsRequest | 1,083,188 | 1,083,032 | -156 | -0.01%
Sema.ConformanceAccessScopeRequest | 270,745 | 270,904 | 159 | 0.06%
Sema.ConformanceHasEffectRequest | 28 | 28 | 0 | 0.0%
Sema.ConstantValueInfoRequest | 1 | 1 | 0 | 0.0%
Sema.ContinueTargetRequest | 2,740 | 2,741 | 1 | 0.04%
Sema.CursorInfoRequest | 0 | 0 | 0 | 0.0%
Sema.CustomAttrNominalRequest | 31,384 | 31,352 | -32 | -0.1%
Sema.CustomAttrTypeRequest | 5,002 | 5,002 | 0 | 0.0%
Sema.CustomRefCountingOperation | 0 | 0 | 0 | 0.0%
Sema.CxxRecordAsSwiftType | 0 | 0 | 0 | 0.0%
Sema.CxxRecordSemantics | 69,769 | 69,599 | -169 | -0.24%
Sema.DefaultAndMaxAccessLevelRequest | 141,395 | 141,459 | 64 | 0.05%
Sema.DefaultArgumentExprRequest | 173,674 | 173,822 | 148 | 0.09%
Sema.DefaultArgumentInitContextRequest | 925 | 927 | 2 | 0.27%
Sema.DefaultArgumentTypeRequest | 17,822 | 17,858 | 35 | 0.2%
Sema.DefaultDefinitionTypeRequest | 4,020 | 4,023 | 3 | 0.07%
Sema.DefaultInitializerIsolation | 1,563,887 | 1,565,081 | 1,193 | 0.08%
Sema.DefaultTypeRequest | 819,226 | 820,391 | 1,165 | 0.14%
Sema.DerivativeAttrOriginalDeclRequest | 0 | 0 | 0 | 0.0%
Sema.DifferentiableAttributeTypeCheckRequest | 0 | 0 | 0 | 0.0%
Sema.DirectLookupRequest | 103,792,087 | 103,814,268 | 22,181 | 0.02%
Sema.DirectOperatorLookupRequest | 2,844,443 | 2,846,859 | 2,416 | 0.08%
Sema.DirectPrecedenceGroupLookupRequest | 1,047,114 | 1,048,242 | 1,128 | 0.11%
Sema.DistributedModuleIsAvailableRequest | 67 | 67 | 0 | 0.0%
Sema.DynamicallyReplacedDeclRequest | 46,976,158 | 46,945,136 | -31,022 | -0.07%
Sema.EnumElementExprPatternRequest | 29 | 29 | 0 | 0.0%
Sema.EnumRawTypeRequest | 52,480 | 52,539 | 59 | 0.11%
Sema.EnumRawValuesRequest | 31,420 | 31,433 | 13 | 0.04%
Sema.ExistentialConformsToSelfRequest | 36,810 | 36,831 | 21 | 0.06%
Sema.ExpandAccessorMacros | 1,261,440 | 1,262,855 | 1,414 | 0.11%
Sema.ExpandBodyMacroRequest | 537,143 | 537,871 | 728 | 0.14%
Sema.ExpandChildTypeRefinementContextsRequest | 890,979 | 891,525 | 546 | 0.06%
Sema.ExpandExtensionMacros | 1,290,105 | 1,288,947 | -1,158 | -0.09%
Sema.ExpandMacroExpansionDeclRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandMacroExpansionExprRequest | 678 | 678 | 0 | 0.0%
Sema.ExpandMemberAttributeMacros | 84,573,926 | 84,423,966 | -149,960 | -0.18%
Sema.ExpandPeerMacroRequest | 61,652,975 | 61,614,507 | -38,468 | -0.06%
Sema.ExpandPreambleMacroRequest | 537,143 | 537,871 | 728 | 0.14%
Sema.ExpandSynthesizedMemberMacroRequest | 369,269 | 369,510 | 240 | 0.07%
Sema.ExplicitCaughtTypeRequest | 38,260 | 38,353 | 93 | 0.24%
Sema.ExprPatternMatchRequest | 114,485 | 114,490 | 5 | 0.0%
Sema.ExtendedNominalRequest | 829,769 | 830,743 | 973 | 0.12%
Sema.ExtendedTypeRequest | 259,297 | 259,659 | 362 | 0.14%
Sema.ExternalMacroDefinitionRequest | 86 | 86 | 0 | 0.0%
Sema.FragileFunctionKindRequest | 3,583,923 | 3,586,679 | 2,756 | 0.08%
Sema.FunctionOperatorRequest | 19,628 | 19,699 | 70 | 0.36%
Sema.GenericParamListRequest | 8,723,832 | 8,682,102 | -41,730 | -0.48%
Sema.GenericSignatureRequest | 4,215,198 | 4,179,707 | -35,491 | -0.84%
Sema.GetDestructorRequest | 48,320 | 48,380 | 60 | 0.13%
Sema.GetDistributedActorArgumentDecodingMethodRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorIDPropertyRequest | 72 | 72 | 0 | 0.0%
Sema.GetDistributedActorImplicitCodableRequest | 117 | 117 | 0 | 0.0%
Sema.GetDistributedActorInvocationDecoderRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorSystemPropertyRequest | 63 | 63 | 0 | 0.0%
Sema.GetDistributedActorSystemRemoteCallFunctionRequest | 40 | 40 | 0 | 0.0%
Sema.GetDistributedRemoteCallArgumentInitFunctionRequest | 14 | 14 | 0 | 0.0%
Sema.GetDistributedRemoteCallTargetInitFunctionRequest | 14 | 14 | 0 | 0.0%
Sema.GetDistributedTargetInvocationDecoderDecodeNextArgumentFunctionRequest | 21 | 21 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordArgumentFunctionRequest | 16 | 16 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordErrorTypeFunctionRequest | 8 | 8 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordReturnTypeFunctionRequest | 13 | 13 | 0 | 0.0%
Sema.GetDistributedTargetInvocationResultHandlerOnReturnFunctionRequest | 12 | 12 | 0 | 0.0%
Sema.GetDistributedThunkRequest | 80 | 80 | 0 | 0.0%
Sema.GetSourceFileAsyncNode | 142 | 142 | 0 | 0.0%
Sema.GlobalActorAttributeRequest | 54,734,085 | 54,697,943 | -36,141 | -0.07%
Sema.GlobalActorInstanceRequest | 3,054 | 3,042 | -11 | -0.38%
Sema.HasCircularInheritedProtocolsRequest | 19,916 | 19,931 | 15 | 0.08%
Sema.HasCircularRawValueRequest | 33,915 | 33,924 | 8 | 0.03%
Sema.HasDefaultInitRequest | 131,033 | 131,107 | 73 | 0.06%
Sema.HasDynamicCallableAttributeRequest | 37,140 | 37,154 | 14 | 0.04%
Sema.HasDynamicMemberLookupAttributeRequest | 664,843 | 665,311 | 468 | 0.07%
Sema.HasImportsMatchingFlagRequest | 793,011 | 794,241 | 1,230 | 0.16%
Sema.HasInitAccessorRequest | 6,894,809 | 6,892,414 | -2,395 | -0.03%
Sema.HasIsolatedSelfRequest | 61,016,828 | 60,978,495 | -38,333 | -0.06%
Sema.HasMemberwiseInitRequest | 74,717 | 74,736 | 19 | 0.03%
Sema.HasSelfOrAssociatedTypeRequirementsRequest | 17,021 | 17,046 | 24 | 0.14%
Sema.HasStorageRequest | 19,315,885 | 19,301,197 | -14,688 | -0.08%
Sema.HasUserDefinedDesignatedInitRequest | 131,037 | 131,111 | 74 | 0.06%
Sema.IDEInspectionFileRequest | 0 | 0 | 0 | 0.0%
Sema.ImplementsAttrProtocolRequest | 0 | 0 | 0 | 0.0%
Sema.ImplicitKnownProtocolConformanceRequest | 286,093 | 286,389 | 296 | 0.1%
Sema.InferredGenericSignatureRequest | 261,292 | 262,112 | 819 | 0.31%
Sema.InheritedDeclsReferencedRequest | 10,005,971 | 9,952,142 | -53,829 | -0.54%
Sema.InheritedProtocolsRequest | 957,304 | 955,070 | -2,234 | -0.23%
Sema.InheritedTypeRequest | 546,544 | 547,478 | 934 | 0.17%
Sema.InitAccessorPropertiesRequest | 89,344 | 89,390 | 46 | 0.05%
Sema.InitAccessorReferencedVariablesRequest | 0 | 0 | 0 | 0.0%
Sema.InitKindRequest | 256,530 | 256,734 | 204 | 0.08%
Sema.InterfaceTypeRequest | 29,325,616 | 29,254,750 | -70,866 | -0.24%
Sema.InvertibleAnnotationRequest | 1,329,230 | 1,329,013 | -217 | -0.02%
Sema.IsABICompatibleOverrideRequest | 459,623 | 459,780 | 157 | 0.03%
Sema.IsAccessorTransparentRequest | 789,721 | 789,982 | 261 | 0.03%
Sema.IsActorRequest | 3,049,517 | 3,044,365 | -5,152 | -0.17%
Sema.IsCCompatibleFuncDeclRequest | 0 | 0 | 0 | 0.0%
Sema.IsCallAsFunctionNominalRequest | 7,906 | 7,905 | -1 | -0.02%
Sema.IsDeclApplicableRequest | 0 | 0 | 0 | 0.0%
Sema.IsDeclRefinementOfRequest | 78,792 | 78,888 | 96 | 0.12%
Sema.IsDefaultActorRequest | 54,301 | 54,379 | 77 | 0.14%
Sema.IsDistributedActorRequest | 3,295,872 | 3,290,804 | -5,068 | -0.15%
Sema.IsDynamicRequest | 45,487,848 | 45,449,533 | -38,315 | -0.08%
Sema.IsEscapableRequest | 0 | 0 | 0 | 0.0%
Sema.IsFinalRequest | 45,350,501 | 45,313,164 | -37,337 | -0.08%
Sema.IsFunctionBodySkippedRequest | 2,329,129 | 2,330,995 | 1,866 | 0.08%
Sema.IsGetterMutatingRequest | 1,024,284 | 1,025,305 | 1,021 | 0.1%
Sema.IsImplicitlyUnwrappedOptionalRequest | 4,905,249 | 4,895,511 | -9,737 | -0.2%
Sema.IsNonUserModuleRequest | 131,373 | 131,367 | -6 | -0.0%
Sema.IsNoncopyableRequest | 0 | 0 | 0 | 0.0%
Sema.IsObjCRequest | 3,823,469 | 3,826,324 | 2,854 | 0.07%
Sema.IsSafeUseOfCxxDecl | 0 | 0 | 0 | 0.0%
Sema.IsSetterMutatingRequest | 928,466 | 929,074 | 607 | 0.07%
Sema.IsSingleValueStmtRequest | 325,515 | 325,856 | 341 | 0.1%
Sema.IsStaticRequest | 2,001,073 | 2,003,326 | 2,253 | 0.11%
Sema.LocalDiscriminatorsRequest | 720,092 | 720,361 | 269 | 0.04%
Sema.LocalTypeDeclsRequest | 63,956 | 64,007 | 51 | 0.08%
Sema.LookupAllConformancesInContextRequest | 4,066,150 | 4,063,022 | -3,128 | -0.08%
Sema.LookupConformanceInModuleRequest | 81,512,295 | 81,533,788 | 21,492 | 0.03%
Sema.LookupInModuleRequest | 11,811,212 | 11,820,582 | 9,369 | 0.08%
Sema.LookupInfixOperatorRequest | 91,879 | 91,961 | 82 | 0.09%
Sema.LookupIntrinsicRequest | 83 | 83 | 0 | 0.0%
Sema.LookupPostfixOperatorRequest | 74 | 74 | 0 | 0.0%
Sema.LookupPrecedenceGroupRequest | 36,335 | 36,375 | 39 | 0.11%
Sema.LookupPrefixOperatorRequest | 442 | 441 | 0 | -0.11%
Sema.MacroDefinitionRequest | 62 | 62 | 0 | 0.0%
Sema.MangleLocalTypeDeclRequest | 3,264 | 3,264 | 0 | 0.0%
Sema.MemberwiseInitPropertiesRequest | 89,346 | 89,392 | 46 | 0.05%
Sema.ModuleImplicitImportsRequest | 11,731 | 11,735 | 4 | 0.03%
Sema.ModuleLibraryLevelRequest | 22 | 22 | 0 | 0.0%
Sema.ModuleQualifiedLookupRequest | 4,531,519 | 4,532,161 | 642 | 0.01%
Sema.NamedLazyMemberLoadSuccessCount | 75,547,528 | 75,486,908 | -60,620 | -0.08%
Sema.NamingPatternRequest | 307,808 | 308,348 | 540 | 0.18%
Sema.NeedsNewVTableEntryRequest | 1,386,792 | 1,387,935 | 1,143 | 0.08%
Sema.NumAccessorBodiesSynthesized | 442,823 | 443,053 | 230 | 0.05%
Sema.NumAccessorsSynthesized | 690,773 | 691,453 | 680 | 0.1%
Sema.NumConformancePathsRecorded | 1,108,989 | 1,109,024 | 35 | 0.0%
Sema.NumConformancesDeserialized | 7,832,047 | 7,814,782 | -17,265 | -0.22%
Sema.NumConstraintScopes | 116,628,224 | 116,692,391 | 64,167 | 0.06%
Sema.NumConstraintsConsideredForEdgeContraction | 4,010,089 | 4,011,610 | 1,521 | 0.04%
Sema.NumCrossImportsChecked | 0 | 0 | 0 | 0.0%
Sema.NumCrossImportsFound | 0 | 0 | 0 | 0.0%
Sema.NumCyclicOneWayComponentsCollapsed | 0 | 0 | 0 | 0.0%
Sema.NumDeclsDeserialized | 117,812,768 | 117,737,791 | -74,977 | -0.06%
Sema.NumDeclsTypechecked | 4,106,760 | 4,109,223 | 2,463 | 0.06%
Sema.NumLazyIterableDeclContexts | 17,458,645 | 17,434,497 | -24,148 | -0.14%
Sema.NumLazyRequirementSignatures | 1,273,718 | 1,272,332 | -1,386 | -0.11%
Sema.NumLazyRequirementSignaturesLoaded | 904,756 | 904,157 | -598 | -0.07%
Sema.NumLeafScopes | 93,872,607 | 93,923,421 | 50,814 | 0.05%
Sema.NumRequirementMachineCompletionSteps | 31,054,282 | 31,042,694 | -11,588 | -0.04%
Sema.NumRequirementMachineUnifiedConcreteTerms | 1,761,759 | 1,759,197 | -2,562 | -0.15%
Sema.NumRequirementMachines | 2,697,351 | 2,697,188 | -163 | -0.01%
Sema.NumSwift6Errors | 6,051 | 6,050 | -1 | -0.02%
Sema.NumTypesDeserialized | 31,624,017 | 31,610,889 | -13,128 | -0.04%
Sema.NumUnloadedLazyIterableDeclContexts | 13,057,292 | 13,065,674 | 8,382 | 0.06%
Sema.ObjCInterfaceAndImplementationRequest | 2,810,826 | 2,808,984 | -1,842 | -0.07%
Sema.OpaqueReadOwnershipRequest | 635,853 | 636,549 | 696 | 0.11%
Sema.OpaqueResultTypeRequest | 1,015 | 1,015 | 0 | 0.0%
Sema.OperatorPrecedenceGroupRequest | 921 | 923 | 2 | 0.22%
Sema.OverriddenDeclsRequest | 28,519,035 | 28,503,573 | -15,462 | -0.05%
Sema.ParamSpecifierRequest | 2,655,869 | 2,659,072 | 3,203 | 0.12%
Sema.PatternBindingCheckedAndContextualizedInitRequest | 491,501 | 491,813 | 312 | 0.06%
Sema.PatternBindingEntryRequest | 1,012,612 | 1,013,557 | 945 | 0.09%
Sema.PatternTypeRequest | 1,198,342 | 1,199,536 | 1,193 | 0.1%
Sema.PolymorphicEffectKindRequest | 1,293,841 | 1,295,161 | 1,320 | 0.1%
Sema.PolymorphicEffectRequirementsRequest | 24 | 24 | 0 | 0.0%
Sema.PotentialMacroExpansionsInContextRequest | 8,703,052 | 8,705,265 | 2,213 | 0.03%
Sema.PreCheckClosureBodyRequest | 183,536 | 183,779 | 243 | 0.13%
Sema.PreCheckFunctionBodyRequest | 538,965 | 539,612 | 646 | 0.12%
Sema.PreCheckResultBuilderRequest | 1,938 | 1,938 | 0 | 0.0%
Sema.PreCheckReturnStmtRequest | 742,767 | 743,562 | 795 | 0.11%
Sema.PrimaryAssociatedTypesRequest | 84,424 | 84,470 | 46 | 0.06%
…